license notice test.pl
Add license notice/summary to head of test.pl script. The notice has a space for the script description and author name, and contact details, that can be inserted. 
If all that someone downloads is a copy of the script, there is no ambiguity, of authorship, contact details or license.